+++ R/zzz.R
@@ -1,6 +1,5 @@
-## set default options for afex_options:
 .onLoad <- function(libname, pkgname) {
-
+  suppressMessages(bmm_options(reset_options = TRUE))
 }
 
 .onAttach <- function(libname, pkgname) {
@@ -21,8 +20,11 @@
 
   startUpMsg <- c(
     paste0("A short introduction to package is available by calling help(\"bmm\"). \n",
-           "More detailed articles on how to fit different models are available via vignettes(\"bmm\").")
+           "More detailed articles on how to fit different models are available via vignettes(\"bmm\").\n",
+           "You can view the list of currently available models by calling supported_models().\n")
   )
+
+  optionsMsg <- tryCatch2(bmm_options())$message
 
   if (interactive()) {
     if (length(behind_cran) > 0 && behind_cran) {
@@ -32,7 +34,7 @@
         utils::update.packages("bmm")
       }
     } else {
-      packageStartupMessage(banner, versionMsg, startUpMsg)
+      packageStartupMessage(banner, versionMsg, startUpMsg, optionsMsg)
     }
   }
 }